Add Util#double_quotes_required?, #double_quotes_acceptable?, #to_string_literal
RC contained several implementations of similar methods. Not only is the
duplication gone, but the new implementations are more robust. For example,
the old implementation would say that this string requires double quotes:

    "\""

The new implementation will also catch strings which were written with escapes
for no good reason -- like "\x68\x65\x6c\x6c\x6f".

(Oh, if you have trouble reading ASCII literals in hex... that's "hello".)

Along with the introduction of these methods, a problem in
Style/UnneededPercentQ has been fixed. Previously, if a %q() string contained
double quotes and what appeared to be an escape, it would not be flagged for
conversion to a single-quoted string. However, there is no such thing as escapes
in a %q() string. (For example, "\t" is a literal backslash and literal "t".)

Now, a %Q() string with double quotes and an escape will not be flagged...
UNLESS the escape was totally unnecessary, like the example above. A %q() string
with double quotes and what only appears to be an escape WILL be flagged.