[Fix #6067] Prevent auto-correct error for `Performance/InefficientHashSearch`
Fixes #6067.

This PR prevents an auto-correct error for `Performance/InefficientHashSearch`
when a method defined by itself and `include?` method are method chaining.

Essentially, static analysis targeted by RuboCop does not know the type
determined at run-time.

So, this implementation may be false negative if it defines method itself.

```ruby
def keys
  # When delegation instead of an array,
  # it does not know the type until run-time.
  []
end

def my_include?(key)
  keys.include?(key)
end
```

If the above `keys` method defined by itself will return a Hash,
it will be false negative, so there is room for improvement in the future.

The aim of this change is to make it more valuable to prevent auto-correct
error than false negative.