Fix a false negative for `Performance/RegexpMatch`
Follow up https://github.com/rubocop-hq/rubocop/pull/7081.

This PR fixes a false negartive for `Performance/RegexpMatch` when using
RuboCop 0.71 or higher.

This causes CI to fail.
https://circleci.com/gh/rubocop-hq/rubocop-performance/588

This is a regression similar to #47. This PR tries to fix it using
a node pattern that has been used for a long time.